refactor: organized packages
