+++ src/main/java/com/cys4/sensitivediscoverer/RegexSeeder.java
@@ -7,6 +7,7 @@ package com.cys4.sensitivediscoverer;
 import com.cys4.sensitivediscoverer.model.JsonRegexEntity;
 import com.cys4.sensitivediscoverer.model.ProxyItemSection;
 import com.cys4.sensitivediscoverer.model.RegexEntity;
+import com.cys4.sensitivediscoverer.utils.Utils;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
