+++ src/main/java/com/cys4/sensitivediscoverer/event/OptionsScannerUpdateNumThreadsListener.java
@@ -2,7 +2,7 @@
 Copyright (C) 2023 CYS4 Srl
 See the file 'LICENSE' for copying permission
 */
-package com.cys4.sensitivediscoverer;
+package com.cys4.sensitivediscoverer.event;
 
 import com.cys4.sensitivediscoverer.model.ScannerOptions;
