+++ src/main/java/com/cys4/sensitivediscoverer/ui/LogsTableContextMenu.java
@@ -2,7 +2,7 @@
 Copyright (C) 2023 CYS4 Srl
 See the file 'LICENSE' for copying permission
 */
-package com.cys4.sensitivediscoverer.component;
+package com.cys4.sensitivediscoverer.ui;
 
 import burp.api.montoya.MontoyaApi;
 import burp.api.montoya.core.Annotations;
