+++ src/main/java/com/cys4/sensitivediscoverer/ui/PopupMenuButton.java
@@ -2,7 +2,7 @@
 Copyright (C) 2023 CYS4 Srl
 See the file 'LICENSE' for copying permission
 */
-package com.cys4.sensitivediscoverer.component;
+package com.cys4.sensitivediscoverer.ui;
 
 import javax.swing.*;
 import javax.swing.event.PopupMenuEvent;
