+++ src/main/java/com/cys4/sensitivediscoverer/ui/tab/AboutTab.java
@@ -2,9 +2,9 @@
 Copyright (C) 2023 CYS4 Srl
 See the file 'LICENSE' for copying permission
 */
-package com.cys4.sensitivediscoverer.tab;
+package com.cys4.sensitivediscoverer.ui.tab;
 
-import com.cys4.sensitivediscoverer.Utils;
+import com.cys4.sensitivediscoverer.utils.Utils;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
