+++ src/main/java/com/cys4/sensitivediscoverer/ui/tab/ApplicationTab.java
@@ -2,7 +2,7 @@
 Copyright (C) 2023 CYS4 Srl
 See the file 'LICENSE' for copying permission
 */
-package com.cys4.sensitivediscoverer.tab;
+package com.cys4.sensitivediscoverer.ui.tab;
 
 import javax.swing.*;
