+++ src/main/java/com/cys4/sensitivediscoverer/ui/tab/OptionsTab.java
@@ -2,9 +2,12 @@
 Copyright (C) 2023 CYS4 Srl
 See the file 'LICENSE' for copying permission
 */
-package com.cys4.sensitivediscoverer.tab;
+package com.cys4.sensitivediscoverer.ui.tab;
 
 import com.cys4.sensitivediscoverer.*;
+import com.cys4.sensitivediscoverer.event.OptionsScannerUpdateListener;
+import com.cys4.sensitivediscoverer.event.OptionsScannerUpdateMaxSizeListener;
+import com.cys4.sensitivediscoverer.event.OptionsScannerUpdateNumThreadsListener;
 import com.cys4.sensitivediscoverer.model.ScannerOptions;
 import com.cys4.sensitivediscoverer.model.UIOptions;
